Fix camera-traveler synchronization
Bug description:
Camera must wait for traveler to recieve the signal so it can take a picture of the board.
To solve this a temporary fix has been applied - buffering the channel with value of 1000.

Fix:
Create a channel for the timer and start the timer as a goroutine from inside travSend goroutine.
Add cascading select statements so that the priority of recieving signals is good. (Otherwise, if 2 channels
get a notification at the same time, GO will choose randomly which block of code to execute).